feat: add graphql
